turbo y float ok, estamos arreglando los nuevbvos .h

+++ INC/player.h
@@ -4,13 +4,13 @@
 # include "t_coordinate.h"
 # include "t_player_movements.h"
 
-typedef struct	s_player
+typedef struct			s_player
 {
 	//t_vector	pos;
-	t_coor	pos;
-	t_coor	end;
-	int		vision_angle;
-	t_player_movements movement;
-}				t_player;
+	t_coor				pos;
+	t_coor				end;
+	float				vision_angle;
+	t_player_movements	movement;
+}						t_player;
 
 #endif
